LineageOS.API (LineageOS API v0.1.0)

API is an interface to the LineageOS API.

Types

builds()

Specs

builds() :: [LineageOS.Build.t()]

devices()

Specs

devices() :: %{required(String.t()) => builds()}

Functions

download(build)

Specs

download(Build.t()) ::
  :ok | {:error, File.posix()} | {:error, HTTPoison.Response.t()}

download downloads the given build.

getBuilds(device)

Specs

getBuilds(String.t()) :: builds()

getBuilds returns a list of builds for a given device.

getDevices()

Specs

getDevices() :: devices()

getDevices returns a list of devices that are supported by the LineageOS team.

getLatestBuild(deviceName)

Specs

getLatestBuild(String.t()) :: Build.t()

getLatestBuild returns the latest build for a given device.